View post
added a new page so when the user click a post ,it takes them to a new page and can view post.they have the option to comment showing the user name comment and time. i tried to but the css in style but didnt work. i have been trying to solve this issue over 1 hour